Refactor com_cronjobs [9]
- Update namespace TaskOption, TaskOptions (again, oops).
- Update TaskOption, TaskOptions directory (`Scheduler` => `Task`).
