+++ administrator/components/com_scheduler/src/Task/TaskOptions.php
@@ -10,7 +10,7 @@
  *
  */
 
-namespace Joomla\Component\Scheduler\Administrator\Tasks;
+namespace Joomla\Component\Scheduler\Administrator\Task;
 
 // Restrict direct access
 defined('_JEXEC') or die;
